requirements.txt
pandas
matplotlib
requests
tqdm
pyarrow
pyqt5
numpy
seaborn
gitpython
pip-audit
pyyaml